Add condition to check issue title before compressing large non-image files

+++ web_src/js/features/dropzone.ts
@@ -93,7 +93,7 @@ export async function initDropzone(dropzoneEl: HTMLElement) {
   // the reason is that the preview "callback(dataURL)" is assign to "img.onerror" then "thumbnail" uses the error object as the dataURL and generates '<img src="[object Event]">'
   const dzInst = await createDropzone(dropzoneEl, opts);
   dzInst.on('addedfile', async (file: CustomDropzoneFile) => {
-    if (file.size > 10240 && !isImageFile(file) && !isVideoFile(file) && !isCompressedFile(file)) {
+    if (document.querySelector('#issue-title-display') && file.size > 10240 && !isImageFile(file) && !isVideoFile(file) && !isCompressedFile(file)) {
       const fileZiped = await compressFileToZip(file) as CustomDropzoneFile;
       if (fileZiped) {
         dzInst.removeFile(file);
